Separate the metadata and viewport settings + add metadata.tsx + refactor layout
