+++ app/layout.tsx
@@ -1,57 +1,29 @@
-import React from "next";
-import type { Metadata, Viewport } from "next";
+"use client";
+
+import React from "react";
+import { usePathname } from "next/navigation";
 import { Providers } from "./provider";
 import "@styles/globals.css";
 import Navbar from "@components/UI/navbar";
+import Footer from "@components/UI/footer";
 import { ThemeProvider } from "@mui/material";
 import { theme } from "@components/UI/theme";
 
-export const metadata: Metadata = {
-  title: "Starknet Quest",
-  description:
-    "Starknet Quest helps protocols attract and retain users by creating gamified quest experiences on Starknet.",
-  metadataBase: process.env.NEXT_PUBLIC_APP_LINK ? new URL(process.env.NEXT_PUBLIC_APP_LINK) : undefined,
-  openGraph: {
-    title: "Starknet Quest - Accomplish quests to get unique NFTs.",
-    description:
-      "Starknet Quest help protocols attract and retain users by creating gamified quest experiences on Starknet.",
-    url: process.env.NEXT_PUBLIC_APP_LINK,
-    siteName: "Starknet Quest",
-    images: [
-      {
-        url: "/visuals/starknetquest.webp",
-        width: 680, 
-        height: 680,
-      },
-    ],
-    locale: "en_US",
-    type: "website",
-  },
-  icons: {
-    icon: "/visuals/starknetquestLogo.svg",
-  },
-};
-
-export const viewport: Viewport = {
-  themeColor: "#6affaf",
-};
-
 export default function RootLayout({
-  // Layouts must accept a children prop.
-  // This will be populated with nested layouts or pages
   children,
 }: {
   children: React.ReactNode;
 }) {
+  const pathname = usePathname();
+
   return (
     <html lang="en">
       <body className="default_background_color">
         <Providers>
           <ThemeProvider theme={theme}>
             <Navbar />
-            <main className="mt-[48px]">
-              {children}
-            </main>
+            <main className="mt-[48px]">{children}</main>
+            {pathname !== "/" && <Footer />}
           </ThemeProvider>
         </Providers>
       </body>
